Logistic regression predicts full distribution
This is extremely interesting. The original code just tries to make sure
the largest number in the distribution is the same as the correct label.
This code tries to match the whole distribution (get close to 0s for
wrong digits, get close to 1s for correct digits).

What's interesting is
that when you report the *max* difference (over any digit) between
expected an actual, and take the mean of that, you get about twice the
'error' as if you're just trying to classify. This means that picking
the best 'hides' an error, where it's 50% sure that it's a 4, but 100%
sure it's a 9, for example.

Alternatively, you can report the mean,
which gives bonus points for getting close to zero for 9/10 labels,
instead of just throwing away the example if you misclassify.

Interesting stuff.